Move source files to `source` folder
