+++ index.js
@@ -1,6 +1,6 @@
-const create = require('./create');
-const useTailwind = require('./use-tailwind');
-const TailwindProvider = require('./tailwind-provider');
+const create = require('./source/create');
+const useTailwind = require('./source/use-tailwind');
+const TailwindProvider = require('./source/tailwind-provider');
 
 exports.create = create;
 exports.useTailwind = useTailwind;
