+++ source/cli.js
