+++ source/lib/evaluate-style.js
