+++ tests/helpers/compile.js
@@ -1,8 +1,8 @@
 const fs = require('fs');
 const execa = require('execa');
 const tempfile = require('tempfile');
-const build = require('../../build');
-const create = require('../../create');
+const build = require('../../source/build');
+const create = require('../../source/create');
 
 const compile = async (classNames, config, media) => {
 	const input = tempfile();
